Move config file wrangling into its own module.
