+++ CMakeLists.txt
@@ -110,6 +110,7 @@ endif ()
 
 add_executable(stubby
   src/stubby.c
+  src/configfile.c
   src/log.c
   src/util.c
   src/yaml/convert_yaml_to_json.c
